test: split return API tests by scenario
