+++ packages/template-app/__tests__/return-missing-deposit.test.ts
@@ -1,0 +1,23 @@
+import { jest } from "@jest/globals";
+import type { NextRequest } from "next/server";
+import { setupReturnMocks, type SessionSubset } from "./helpers/return";
+
+afterEach(() => jest.resetModules());
+
+describe("/api/return missing deposit", () => {
+  test("returns ok false when deposit missing", async () => {
+    const session: SessionSubset = { metadata: {} } as SessionSubset;
+    const { refundCreate } = setupReturnMocks({ session });
+
+    const { POST } = await import("../src/api/return/route");
+    const res = await POST({
+      json: async () => ({ sessionId: "sess" }),
+    } as unknown as NextRequest);
+
+    expect(refundCreate).not.toHaveBeenCalled();
+    expect(await res.json()).toEqual({
+      ok: false,
+      message: "No deposit found",
+    });
+  });
+});
